requirements.txt
uiautomator2>=3.2.10
paddlepaddle>=3.0.0
paddleocr>=2.6.0
opencv-python>=4.6.0
pillow>=9.0.0
pyyaml>=6.0
loguru>=0.7.0 